Cleanup code and add comments
